Vista Inventario

+++ Temunt/Models/producto.cs
@@ -5,12 +5,11 @@ namespace Temunt.Models
 {
     public class producto
     {
-
         [Key]
         public int id_prod { get; set; }
         public string nombre { get; set; }
         public string descripcion { get; set; }
-        public int precio { get; set; }
+        public decimal precio { get; set; }  // <-- CORREGIDO
         public int stock { get; set; }
         public int id_cat { get; set; }
         public int id_prov { get; set; }
